requirements.txt
requests
oauthlib
requests-oauthlib
python-decouple
pre-commit
